+++ src/xulbux/cli/help.py
@@ -85,7 +85,7 @@ CLI_HELP: Final[StyledText] = StyledText(
     (H.BORDER("  │ "), H.PUNCT("# ", S.ITALIC("LIBRARY CONSTANTS                               ")), H.BORDER("│")),
     (H.BORDER("  │ "), H.IMPORT("from "), H.LIB("xulbux"), (S.DIM | H.LIB)("."), H.LIB("base"), (S.DIM | H.LIB)("."), H.LIB("consts "), H.IMPORT("import "), H.CONST("COLOR"), H.PUNCT(", "), H.CONST("CHARS"), H.PUNCT(", "), H.CONST("ANSI "), H.BORDER("│")),
     (H.BORDER("  │ "), H.PUNCT("# ", S.ITALIC("Main Classes                                    ")), H.BORDER("│")),
-    (H.BORDER("  │ "), H.IMPORT("from "), H.LIB("xulbux "), H.IMPORT("import "), H.CLS("Code"), H.PUNCT(", "), H.CLS("Color"), H.PUNCT(", "), H.CLS("Console"), H.PUNCT(", "), H.META("...      "), H.BORDER("│")),
+    (H.BORDER("  │ "), H.IMPORT("from "), H.LIB("xulbux "), H.IMPORT("import "), H.CLS("code"), H.PUNCT(", "), H.CLS("color"), H.PUNCT(", "), H.CLS("console"), H.PUNCT(", "), H.META("...      "), H.BORDER("│")),
     (H.BORDER("  │ "), H.PUNCT("# ", S.ITALIC("module specific imports                         ")), H.BORDER("│")),
     (H.BORDER("  │ "), H.IMPORT("from "), H.LIB("xulbux"), (S.DIM | H.LIB)("."), H.LIB("color "), H.IMPORT("import "), H.FN("rgba"), H.PUNCT(", "), H.FN("hsla"), H.PUNCT(", "), H.FN("hexa         "), H.BORDER("│")),
     H.BORDER("  ╰───────────────────────────────────────────────────╯"),
@@ -95,6 +95,7 @@ CLI_HELP: Final[StyledText] = StyledText(
     (H.BORDER("  │ "), (S.BR.BLUE | S.link("https://github.com/xulbux/python-lib-xulbux/wiki"))("github.com/xulbux/python-lib-xulbux/wiki"), "          ", H.BORDER("│")),
     H.BORDER("  ╰───────────────────────────────────────────────────╯"),
     "",
+    sep="\n",
 )
 # fmt: on